Refactor SmartAnswerPresenter#questions
Following princliple "Tell, don't ask" I've encapsulated the implementation
of this method within corresponding classes.
All question presenters inherit from the QuestionPresenter, so behaviour for
them does not change, they return themselves as a single member of an array.
For all other presenters the return value will remain an empty array,
implemented on the superclass of them all - NodePresenter.

This refactoring is in preparation of adding support for multiple questions
per node.