Add Universal Credit to the benefits list for energy grants Smart Answer
As of 2015 April the new Universal Credit replaces the following:
- Jobseeker's Allowance
- Housing Benefit
- Working Tax Credit
- Child Tax Credit
- Employment and Support Allowance
- Income Support

This adds Universal Credit as an option, however it is still too early to remove
the other options as there are still users claiming those older types of credits.

If user selects Universal Credits, the behaviour is exactly the same as for
the process when the user receives Jobseeker's allowance, but the outcomes must
have an extra paragraph about Affordable Warmth Obligation.